+++ src/main/java/log/charter/gui/components/preview3D/data/ChordBoxDrawData.java
@@ -3,13 +3,13 @@ package log.charter.gui.components.preview3D.data;
 import log.charter.data.song.enums.Mute;
 
 public class ChordBoxDrawData {
-	public final int originalPosition;
-	public final int position;
+	public final double originalPosition;
+	public final double position;
 	public final Mute mute;
 	public final boolean onlyBox;
 	public final boolean withTop;
 
-	public ChordBoxDrawData(final int position, final Mute mute, final boolean onlyBox, final boolean withTop) {
+	public ChordBoxDrawData(final double position, final Mute mute, final boolean onlyBox, final boolean withTop) {
 		originalPosition = position;
 		this.position = position;
 		this.mute = mute;
@@ -17,7 +17,7 @@ public class ChordBoxDrawData {
 		this.withTop = withTop;
 	}
 
-	public ChordBoxDrawData(final int position, final ChordBoxDrawData other) {
+	public ChordBoxDrawData(final double position, final ChordBoxDrawData other) {
 		originalPosition = other.originalPosition;
 		this.position = position;
 		mute = other.mute;
