+++ src/main/java/log/charter/gui/components/preview3D/data/NoteDrawData.java
@@ -13,8 +13,8 @@ import log.charter.data.song.notes.ChordNote;
 import log.charter.data.song.notes.Note;
 
 public class NoteDrawData {
-	public final int position;
-	public final int endPosition;
+	public final double position;
+	public final double endPosition;
 	public final int string;
 	public final int fret;
 	public final boolean accent;
@@ -29,13 +29,13 @@ public class NoteDrawData {
 	public final boolean tremolo;
 	public final boolean linkPrevious;
 
-	public final int originalPosition;
-	public final int trueLength;
+	public final double originalPosition;
+	public final double trueLength;
 	public final boolean withoutHead;
 	public final boolean isChordNote;
 
-	public NoteDrawData(final int notePosition, final int noteEndPosition, final int minPosition, final int maxPosition,
-			final Note note, final boolean linkPrevious) {
+	public NoteDrawData(final double notePosition, final double noteEndPosition, final double minPosition,
+			final double maxPosition, final Note note, final boolean linkPrevious) {
 		originalPosition = notePosition;
 		trueLength = noteEndPosition - notePosition;
 
@@ -68,8 +68,8 @@ public class NoteDrawData {
 				: 0;
 	}
 
-	public NoteDrawData(final int chordPosition, final int chordNoteEndPosition, final int minPosition,
-			final int maxPosition, final Chord chord, final int string, final int fret, final ChordNote chordNote,
+	public NoteDrawData(final double chordPosition, final double chordNoteEndPosition, final double minPosition,
+			final double maxPosition, final Chord chord, final int string, final int fret, final ChordNote chordNote,
 			final boolean linkPrevious, final boolean shouldHaveLength) {
 		originalPosition = chordPosition;
 		trueLength = shouldHaveLength ? chordNoteEndPosition - chordPosition : 0;
@@ -102,7 +102,8 @@ public class NoteDrawData {
 				: 0;
 	}
 
-	public NoteDrawData(final int truePosition, final int position, final int endPosition, final NoteDrawData note) {
+	public NoteDrawData(final double truePosition, final double position, final double endPosition,
+			final NoteDrawData note) {
 		originalPosition = truePosition;
 		trueLength = note.trueLength;
