+++ src/main/java/log/charter/gui/components/preview3D/drawers/Preview3DHandShapesDrawer.java
@@ -60,8 +60,8 @@ public class Preview3DHandShapesDrawer {
 		final double y = getChartboardYPosition(data.currentStrings()) + 0.0002;
 		final boolean arpeggio = handShape.template.arpeggio;
 
-		final int timeFrom = max(0, handShape.timeFrom - drawData.time);
-		final int timeTo = min(getVisibility(), handShape.timeTo - drawData.time);
+		final double timeFrom = max(0, handShape.timeFrom - drawData.time);
+		final double timeTo = min(getVisibility(), handShape.timeTo - drawData.time);
 		if (timeTo < 0) {
 			return;
 		}
